# Minor performance improvements, code cleanup and very local fixes