--- a/lib/permutat.g
+++ b/lib/permutat.g
@@ -780,6 +780,7 @@ InstallMethod( Order,
     [ IsPerm ],
     ORDER_PERM );
 
+
 #############################################################################
 ##
 #O  DistancePerms( <perm1>, <perm2> ) . returns NrMovedPoints( <perm1>/<perm2> )