--- a/tst/testbugfix/2018-09-05-CacheGL.tst
+++ b/tst/testbugfix/2018-09-05-CacheGL.tst
@@ -1,0 +1,5 @@
+gap> g:=SL(4,3);;
+gap> NiceMonomorphism(g);;
+gap> g:=GL(4,3);;
+gap> hom:=NiceMonomorphism(g);;
+gap> List(GeneratorsOfGroup(g),x->Image(hom,x));;