Most of this is changes that are entirely "under the hood" for a user and improve performance in particular (often too large to allow for specific test examples) cases I observed, or that clean up ugly code. It is not worth mentioning it specifically in a release announcement.

The utility function `MaxesAlmostSimple` is made an operation to allow for future packages to overload (and thus improve the maximal subgroups code).

The only user-visible changes are:

- clarifications in how to interpret the output of `SubgroupsSolvableGroup`.
- A (prior forgotten) warning about a calculation using a classification from the primitive groups library.
- a (very local) fix for #2724 .