@hulpke 2018-08-30
Dear @fingolfin 

> Any review remark I make of course just constitutes my personal opinion, and you are of course free to ignore it.

Well, since proposed changes cannot be merged in without an approving review, any review that comments but does not approve is implicitly stating that these are issues without which the proposed contribution should be rejected.

> Anyway, all in all, over the past months I got a strong feeling that my remarks on your PRs are perceived more as nuisance than useful, and as "wasting your time".

I know that I do not react well to stylistic suggestions of changes. I'm sorry.
In fact, for changes like these (which are cleanup and bug fixes) I am not soliciting for reviews but for the fact that I am blocked from merging without an approving review.

> Therefore, in order to not waste both our time, I will refrain from reviewing your PRs for the time being, and leave it to others to take care of that.

Then, why do we have the enforced review for every contribution? I fear that we have technical features of github dictate the way how we collaborate.