@fingolfin 2018-08-30
> > Any review remark I make of course just constitutes my personal opinion, and you are of course free to ignore it.
> 
> Well, since proposed changes cannot be merged in without an approving review, any review that comments but does not approve is implicitly stating that these are issues without which the proposed contribution should be rejected.

I disagree with this interpretation of what a "review that comments but does not approve" means. It would ultimately imply that one must not comment on a PR, ever, unless to either approve it or reject it. So, if I spot a typo, I shouldn't point it out, unless I am willing and able to review the whole PR? That makes no sense to me. Esp. bigger PRs are hard to review in one go, so a partial review sometimes is the best I can afford with limited time and energy.

In any case, this interpretation does not match my intentions at all. Indeed, if I reject a PR in its current form, I use the "request changes" option. If I think a PR could be merged in its current form (even if there are some changes I'd prefer to see), I'll approve it. If I neither "approve" nor "request changes", that can mean a lot of things: that I did not have time to review it all; that I do not feel informed enough to express an opinion either way; that I am conflicted about it; or something completely else. And based on past discussions with other GAP developers, many (most?) seem to have a rather similar view on this. But they can and should speak for themselves, of course.

In the case of this particular PR, it's really that I did not have the energy to slog through all of it. There are tons of different, unrelated changes mingled up together in it. IMHO, if instead of this, it was split into several smaller commits, somebody likely would have reviewed most or even all of them already. As it is, I'd have to either review all of them (for which I don't have the energy, and quite frankly, also don't see why I should spend the effort to accommodate you, when I'd tell anybody else to split up the PR into multiple). Or I could either blindly rubber stamp it (but I don't like doing that); or request that you break it up (but based on past experience, I don't expect you to react favorably to this, so I didn't bother to do so). Nor do I want to reject it, as I have no actual concerns about it (by virtue of not having reviewed it).

Thus, I am instead leaving it for somebody else to review.


> 
> > Anyway, all in all, over the past months I got a strong feeling that my remarks on your PRs are perceived more as nuisance than useful, and as "wasting your time".
> 
> I know that I do not react well to stylistic suggestions of changes. I'm sorry.
> In fact, for changes like these (which are cleanup and bug fixes) I am not soliciting for reviews but for the fact that I am blocked from merging without an approving review.

Well, here we have a fundamental disagreement then. In my view, every change should be reviewed by at least one other party. From that view point, it does not matter if you are soliciting for reviews: any changes that goes in is reviewed, period. To avoid a misunderstanding, let me emphasize that I expressed that this is simply *my personal view* (although I believe that several other GAP developers share it), and that it's not me being a dictator blocking your PR out of a whim. You can raise this on the mailing list if you like, perhaps the GAP team can agree on a different approach.


> 
> > Therefore, in order to not waste both our time, I will refrain from reviewing your PRs for the time being, and leave it to others to take care of that.
> 
> Then, why do we have the enforced review for every contribution? I fear that we have technical features of github dictate the way how we collaborate.

This is not a github dictate; it's a setting we chose to activate. And which we can of course deactivate again, should we (as a collective) decide to do so. Feel free to bring this up on one of the mailing lists for discussion, too!

Here's my view on it, though:

We enforce reviews of every contribution to ensure a certain minimal quality level for them. E.g. that they are understandable, sufficiently explained/documented, accompanied by sensible tests, etc. To be frank, I usually have (mild!) concerns about that with your PRs (e.g. lack of tests -- and I am not talking about tests for performance improvements, where you explain that the tests you have take hours or days -- that's of course perfectly reasonable!; I am talking about tests that verify that e.g. a bug fix does actually fix a bug; or tests that verify that a given change does not break existing functionality). That said, so far I usually was happy to give you far more leeway than I'd give almost any other contributor with my reviews, because I of course know you are a master of your domain, and you make very valuable contributions.

However, I learned that you do not appreciate the feedback I give on that work, and even complained to others about it (indirectly at least). Very well. I could write a lot about that, but in the end, I'll just accept it for the time being, and move on, and, as I wrote, not waste both our time with it then.

But it's not my intention to block your PRs with that! Rather, I expect others to review them. There is no magical law that says that I am the only one who can review your PRs.
If I may suggest, you can do what I often do if I don't get a review on a PR quickly: I specifically request reviews from other people (Chris, Markus, Steve, Frank, Thomas, ...), via the GitHub UI for that -- I am hopeful that they will either review your PR, or explain why they won't.

That said, we do indeed have a shortage of people who actively review PRs. That is a problem. And perhaps it means we should stop requiring reviews again (although personally I'd consider it a shame). Once again, this is of course something that can be discussed, ideally on the mailing list.