@alex-konovalov 2018-09-07
For the record, I've compared performance in this PR and in nightly builds yesterday.

1. teststandard

Linux:
* master: total   2914925 ms (799008 ms GC) and 490GB allocated
* this PR: total   2774530 ms (782115 ms GC) and 506GB allocated

macOS:
* master: total   5174297 ms (1372988 ms GC) and 446GB allocated
* this PR: total   5288641 ms (1438456 ms GC) and 467GB allocated

2. testinstall

Linux:
* master: total     86346 ms (43542 ms GC) and 10.5GB allocated
* this PR: total     91369 ms (45655 ms GC) and 10.4GB allocated

macOS:
* master: total    174664 ms (82735 ms GC) and 10.5GB allocated 
* this PR: total    168886 ms (80093 ms GC) and 10.4GB allocated